// src/api/sf/data-contracts.ts
export interface ObjectStorageResourceModel {
  /**
   * 本地总空间
   * @format int64
   */
  allocatedSpace?: number

  /** 访问存储资源的密码 */
  password?: string

  /** performanceType,如：HIGH,MIDDLE,LOW */
  performanceType?: "PERFOR_NULL" | "HIGH" | "MIDDLE" | "LOW"

  /**
   * 存储id
   * @format int64
   */
  storageId?: number

  /** 存储名称 */
  storageName?: string

  /** 存储状态,如：STORAGE_BROKEN,STORAGE_ONLINE,STORAGE_OFFLINE,CONFIG_MISMATCH */
  storageStatus?: "STORAGE_NULL" | "STORAGE_BROKEN" | "STORAGE_ONLINE" | "STORAGE_OFFLINE" | "CONFIG_MISMATCH"

  /** 存储类型,如：NAS,DBS,REMOTE_DISK,LOCAL_DISK,DOS */
  storageType?: "NULL" | "S3" | "DOS" | "MIO" | "NAS" | "DBS" | "REMOTE_DISK" | "LOCAL_DISK"

  /**
   * 总空间
   * @format int64
   */
  totalSpace?: number

  /**
   * 存储资源地址(对象存储地址或文件存储挂载点)
   * @example 127.0.0.1:8080
   */
  url?: string

  /**
   * 使用空间
   * @format int64
   */
  usedSpace?: number

  /** 访问存储资源的用户名 */
  userName?: string
}
// ...
export interface ResponseModelObject {
  /**
   * 响应码
   * @format int32
   * @example 0
   */
  code?: number

  /**
   * 响应码描述信息
   * @example Success
   */
  description?: string

  /** 响应数据 */
  data?: object
}